Move load image into frdc_dataset
